# xs2a/xs2a-api/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Reference Implementation of XS2A REST Services",
    "version" : "0.9.2-SNAPSHOT",
    "title" : "XS2A Services",
    "contact" : {
      "name" : "Andreas Reetz",
      "email" : "andreas.reetz@sdv-it.de"
    }
  },
  "host" : "api.sdv-it.de",
  "basePath" : "/tpp",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/{version}/consents" : {
      "post" : {
        "tags" : [ "Consents" ],
        "summary" : "Account Information Consent Request (page 48) - creates an account information consent resource at the ASPSP regarding access to specified accounts",
        "description" : "",
        "operationId" : "informationConsent",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-Corporate-ID",
          "in" : "header",
          "description" : "Only used in a corporate context",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Request body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/InformationConsentRequest"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "OK, request received",
            "schema" : {
              "$ref" : "#/definitions/InformationConsentResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      }
    },
    "/{version}/consents/{consent-id}" : {
      "get" : {
        "tags" : [ "Consents" ],
        "summary" : "Get Consent Request (page 56) - returns the content of of an account information consent object",
        "description" : "",
        "operationId" : "consentRequest",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "consent-id",
          "in" : "path",
          "description" : "Consent ID as received from account information consent",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-Corporate-ID",
          "in" : "header",
          "description" : "Only used in a corporate context",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, request received",
            "schema" : {
              "$ref" : "#/definitions/ConsentRequestResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      },
      "put" : {
        "tags" : [ "Consents" ],
        "summary" : "Update PSU Data (page 70) - updates the account information consent data by PSU data",
        "description" : "",
        "operationId" : "updatePSUData",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "consent-id",
          "in" : "path",
          "description" : "Consent ID as received from account information consent",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-Corporate-ID",
          "in" : "header",
          "description" : "Only in corporate context",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Request body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/UpdatePSUDataRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, request received",
            "schema" : {
              "$ref" : "#/definitions/UpdatePSUDataResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      },
      "delete" : {
        "tags" : [ "Consents" ],
        "summary" : "Delete an Account Information Consent Object (page 58) - deletes the consent object",
        "description" : "",
        "operationId" : "deleteConsent",
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "consent-id",
          "in" : "path",
          "description" : "Consent ID as received from account information consent",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "OK, request received"
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      }
    },
    "/{version}/consents/{consent-id}/status" : {
      "get" : {
        "tags" : [ "Consents" ],
        "summary" : "Get Status Request (page 55) - reads the transaction status of the addressed consent resource",
        "description" : "",
        "operationId" : "statusRequest",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "consent-id",
          "in" : "path",
          "description" : "Consent ID as received from account information consent",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-Corporate-ID",
          "in" : "header",
          "description" : "Only used in a corporate context",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, request received",
            "schema" : {
              "$ref" : "#/definitions/StatusResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      }
    },
...
    "/{version}/payments/{payment-product}/{payment-id}" : {
      "get" : {
        "tags" : [ "Payments" ],
        "summary" : "Read Payment Data (page ?) - reads the details of an initiated payment",
        "description" : "",
        "operationId" : "readData",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "payment-product",
          "in" : "path",
          "description" : "sepa-credit-transfers / instant-sepa-credit-transfers / target-2-payments / cross-border-credit-transfers",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "payment-id",
          "in" : "path",
          "description" : "Payment ID as received from the Payment Initiation",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, request received",
            "schema" : {
              "$ref" : "#/definitions/PaymentInitiationResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      },
      "put" : {
        "tags" : [ "Payments" ],
        "summary" : "Update PSU data (page 70) - updates the payment initiation data by PSU data",
        "description" : "",
        "operationId" : "updatePSUData",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "payment-product",
          "in" : "path",
          "description" : "sepa-credit-transfers / instant-sepa-credit-transfers / target-2-payments / cross-border-credit-transfers",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "payment-id",
          "in" : "path",
          "description" : "Payment ID as received from the Payment Initiation",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-Corporate-ID",
          "in" : "header",
          "description" : "Only in corporate context",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Request body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/UpdatePSUDataRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, request received",
            "schema" : {
              "$ref" : "#/definitions/UpdatePSUDataResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "InformationConsentRequest" : {
      "type" : "object",
      "required" : [ "access_accounts", "combined_service_indicator", "frequency_per_day", "recurring_indicator", "valid_until" ],
      "properties" : {
        "access_accounts" : {
          "type" : "array",
          "description" : "List of accounts to be accessed",
          "items" : {
            "$ref" : "#/definitions/ConsentAccessAccount"
          }
        },
        "recurring_indicator" : {
          "type" : "boolean",
          "example" : true,
          "description" : "\"true\" if the consent is for requring access"
        },
        "valid_until" : {
          "type" : "string",
          "example" : "2017-10-30",
          "description" : "Valid until date for the requested consent"
        },
        "frequency_per_day" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 4,
          "description" : "Maximum frequency for an access per day"
        },
        "combined_service_indicator" : {
          "type" : "boolean",
          "example" : false,
          "description" : "\"true\": a payment initiation service will be address in the same \"session\""
        }
      }
    },
    "InformationConsentResponse" : {
      "type" : "object",
      "required" : [ "_links" ],
      "properties" : {
        "sca_methods" : {
          "type" : "array",
          "description" : "name or names of autheticated methods",
          "items" : {
            "$ref" : "#/definitions/AuthenticationMethod"
          }
        },
        "chosen_sca_method" : {
          "description" : "name of chosen method",
          "$ref" : "#/definitions/AuthenticationMethod"
        },
        "sca_chalenge_data" : {
          "description" : "Extended information to chosen_sca_method",
          "$ref" : "#/definitions/AuthenticationChallengeData"
        },
        "transaction_status" : {
          "type" : "string",
          "description" : "Status"
        },
        "psu_message" : {
          "type" : "string",
          "description" : "Text to be displayed to the PSU"
        },
        "_links" : {
          "description" : "Link to (PUT) Update PSU Ddata / (GET) Consent Request",
          "$ref" : "#/definitions/Links"
        }
      }
    },
...
    "UpdatePSUDataRequest" : {
      "type" : "object",
      "properties" : {
        "authentication_method_id" : {
          "type" : "string",
          "description" : "authentication method id for Update PSU Data (Authentication Method)"
        },
        "psu_data" : {
          "description" : "PSU Data for Update PSU Data (Authentication)",
          "$ref" : "#/definitions/PSUData"
        },
        "sca_authentication_data" : {
          "type" : "string",
          "description" : "sca_authentication_data - for Authorise Request"
        }
      },
      "description" : "Update PSU Data Request"
    },
...
  }
}